refactor(utils): Implement threading version of astreamify and asignalize
Reorganizes utils.py for better readability and introduces
streamify and signalize decorators for multi-threading compatibility,
such as with PygameGUI. The task set has been removed due to the
unlimited queue size, eliminating the possibility of a Full exception.
A non-blocking get method has also been implemented.

Additional changes include the removal of unnecessary future annotation
imports in several files, renaming of files in the example directory,
and minor updates to the Bot class documentation to align with other
function updates.